Filtering Middleware Catches
Doesn't apply nav_auth middleware to routes going to any of the Auth Stack routes if you have no token because most likely you are being redirected. If not, you will be next immediately anyways and it would cause two 'unauthorized' dispatches.

+++ app/config/nav-utils.js
@@ -14,11 +14,14 @@ const addListener = createReduxBoundAddListener("root");
 
 // Runs on each dispatch to store. Will only listen to Navigation actions in the switch
 const nav_auth = store => next => async action => {
+    console.log(action)
     // If there was no token
     if ( !await AsyncStorage.getItem('userToken') ) {
+
         switch (action.type) {
             // If you were navigating
             case NavigationActions.navigate.toString():
+                if (action.routeName && ['Login', 'Auth'].indexOf(action.routeName) > -1) return next(action)
                 // Change destination route to Auth
                 action.routeName = 'Auth'
                 // Clear out all of local storage (logout)
@@ -27,8 +30,11 @@ const nav_auth = store => next => async action => {
             break;
             // If you were going back using menu
             case NavigationActions.back.toString():
+                if (action.routeName && ['Login', 'Auth'].indexOf(action.routeName) > -1) return next(action)
                 // Dispatch a route change to the store
-                store.dispatch(NavigationActions.navigate('Auth'))
+                store.dispatch(NavigationActions.navigate({
+                    routeName: 'Auth'
+                }))
                 // Log out
                 store.dispatch({ type: 'unauthorized' });
                 AsyncStorage.clear()
